Change: exclude python 4 from dependencies
Apparently python 4 seems to be considered as an option in creating the
dependency list in poetry now.
So, when python 4 should be supported, all other dependencies must support
it, too, apparently.

Which in turn will break stuff as some dependencies already explicitly
dont support python4 - in our case some dependency of pontos, which in
turn makes pontos requzire python <4.

That again makes updating the poetry lock impossible and lead to errors
described in DEVOPS-1536.

The solution is to make this here depend on python <4, too.